refactors the cache module by reorganizing cache-related files into more clearly defined subdirectories
