+++ be/src/cache/disk_cache/block_cache.h
@@ -16,7 +16,7 @@
 
 #include <atomic>
 
-#include "cache/local_disk_cache_engine.h"
+#include "cache/disk_cache/local_disk_cache_engine.h"
 #include "cache/remote_cache_engine.h"
 #include "common/status.h"
