+++ be/src/cache/disk_space_monitor.h
@@ -20,7 +20,7 @@
 #include <unordered_map>
 
 #include "cache/cache_options.h"
-#include "cache/local_disk_cache_engine.h"
+#include "cache/disk_cache/local_disk_cache_engine.h"
 #include "common/status.h"
 #include "fs/fs.h"
 #include "util/disk_info.h"
