+++ be/src/runtime/runtime_state.h
@@ -44,8 +44,8 @@
 #include <unordered_map>
 #include <vector>
 
-#include "cache/block_cache/block_cache.h"
 #include "cache/datacache_utils.h"
+#include "cache/disk_cache/block_cache.h"
 #include "cctz/time_zone.h"
 #include "common/global_types.h"
 #include "common/object_pool.h"
