+++ be/test/CMakeLists.txt
@@ -408,7 +408,8 @@ set(EXEC_FILES
         ./service/service_be/internal_service_test.cpp
         ./service/staros_worker_test.cpp
         ./cache/datacache_utils_test.cpp
-        ./cache/block_cache/block_cache_hit_rate_counter_test.cpp
+        ./cache/disk_cache/block_cache_hit_rate_counter_test.cpp
+        ./cache/mem_cache/page_cache_test.cpp
         ./cache/peer_cache_test.cpp
         ./util/thrift_rpc_helper_test.cpp
         ./util/thrift_client_test.cpp
@@ -602,9 +603,9 @@ SET(DW_TEST_FILES
 
 if ("${WITH_STARCACHE}" STREQUAL "ON")
     list(APPEND DW_TEST_FILES ./cache/disk_space_monitor_test.cpp)
-    list(APPEND DW_TEST_FILES ./cache/lrucache_engine_test.cpp)
-    list(APPEND DW_TEST_FILES ./cache/starcache_engine_test.cpp)
-    list(APPEND DW_TEST_FILES ./cache/block_cache/block_cache_test.cpp)
+    list(APPEND DW_TEST_FILES ./cache/disk_cache/starcache_engine_test.cpp)
+    list(APPEND DW_TEST_FILES ./cache/disk_cache/block_cache_test.cpp)
+    list(APPEND DW_TEST_FILES ./cache/mem_cache/lrucache_engine_test.cpp)
     list(APPEND DW_TEST_FILES ./io/cache_input_stream_test.cpp)
 endif ()
