+++ be/test/cache/disk_cache/test_cache_utils.h
@@ -17,9 +17,9 @@
 #include <fmt/format.h>
 #include <gtest/gtest.h>
 
-#include "cache/block_cache/block_cache.h"
+#include "cache/disk_cache/block_cache.h"
+#include "cache/disk_cache/starcache_engine.h"
 #include "cache/peer_cache_engine.h"
-#include "cache/starcache_engine.h"
 #include "common/logging.h"
 #include "testutil/assert.h"
