+++ be/test/cache/disk_space_monitor_test.cpp
@@ -20,8 +20,8 @@
 #include <cstring>
 #include <filesystem>
 
-#include "cache/block_cache/block_cache.h"
-#include "cache/block_cache/test_cache_utils.h"
+#include "cache/disk_cache/block_cache.h"
+#include "cache/disk_cache/test_cache_utils.h"
 #include "common/logging.h"
 #include "common/statusor.h"
 #include "fs/fs_util.h"
